requirements.txt
openai
requests
beautifulsoup4
python-dotenv

# Machine Learning
scikit-learn
pandas
numpy
matplotlib

# Vector Memory & Embeddings
faiss-cpu
sentence-transformers

# Web & APIs
serpapi
tavily-python

# Kaggle Datasets
kaggle

# GitHub API
PyGithub

# Backend / API Server
fastapi
uvicorn

# Optional Dev Tools
tqdm
rich

# Database
pymongo
